refactor the models
