+++ src/database.ts
@@ -2,10 +2,7 @@ import dotenv from 'dotenv';
 dotenv.config();
 
 import { Sequelize } from 'sequelize';
-import { define as userDefine } from './models/user.model';
-import { define as hookDefine } from './models/hook.model';
-import { define as panelDefine } from './models/panel.model';
-import { define as panelSetDefine } from './models/panelSet.model';
+import { modelDefiners } from './models'
 
 /**
  * SSL is required for Heroku Postgres
@@ -33,11 +30,7 @@ const sequelize = new Sequelize(
 const setup = async () => {
 
     // define models
-    userDefine(sequelize);
-    hookDefine(sequelize);
-    panelDefine(sequelize);
-    panelSetDefine(sequelize);
-    panelDefine(sequelize);
+    modelDefiners.forEach(define => define(sequelize))
 
     // set up associations
